REFACTOR: Make properties a proper subpackage
