+++ preflibtools/properties/__init__.py
@@ -1,1 +1,6 @@
 from preflibtools.properties.basic import *
+from preflibtools.properties.decorators import *
+from preflibtools.properties.distances import *
+from preflibtools.properties.metrics import *
+from preflibtools.properties.singlecrossing import *
+from preflibtools.properties.singlepeakedness import *
