+++ preflibtools/properties/basic.py
@@ -1,45 +1,6 @@
 """ This module describes several procedures to check for basic procedures of PrefLib instances.
 """
 
-from functools import wraps
-
-
-
-class PreferenceIncompatibleError(Exception):
-    """Exception raised if an instance with a different type of preferences than expected was encountered."""
-
-    pass
-
-
-def requires_preference_type(*dtype):
-    """Decorator to filter instances with wrong preference type."""
-
-    def decorator_fn(fn):
-        @wraps(fn)
-        def wrapper_fn(*fn_args, **fn_kw_args):
-            if fn_args[0].data_type not in dtype:
-                type_str = (("%s, " * (len(dtype) - 1)) + "%s") % dtype
-                raise PreferenceIncompatibleError(
-                    "Only the following types are accepted: %s" % type_str
-                )
-            return fn(*fn_args, **fn_kw_args)
-
-        return wrapper_fn
-
-    return decorator_fn
-
-
-def requires_approval(fn):
-    """Decorator to filter ordinal instances that are not approval."""
-
-    @wraps(fn)
-    def wrapper_fn(*fn_args, **fn_kw_args):
-        if not is_approval(fn_args[0]):
-            raise PreferenceIncompatibleError("Only approval preferences are accepted.")
-        return fn(*fn_args, **fn_kw_args)
-
-    return wrapper_fn
-
 
 def num_alternatives(instance):
     """Returns the number of alternatives of the instance.
