+++ tests/properties/test_metrics.py
@@ -1,5 +1,5 @@
 from preflibtools.instances.preflibinstance import OrdinalInstance
-from preflibtools.properties.metrics import *
+from preflibtools.properties import *
 
 from unittest import TestCase
