+++ tests/properties/test_singlepeakedness.py
@@ -1,5 +1,5 @@
 from preflibtools.instances.preflibinstance import OrdinalInstance
-from preflibtools.properties.singlepeakedness import *
+from preflibtools.properties import *
 
 from unittest import TestCase
